packages/bots/axiom-assist-bot/adk-agents/superpowers/requirements.txt
# Superpowers dependencies
aiohttp>=3.8.0

# ------------------------------
# DevOps Agent Libraries
# ------------------------------

# For CodeReviewPower and LogAnalysisPower
google-generativeai

# For VulnerabilityScanPower (Google open source tool)
osv-scanner

# For CodeReviewPower (to post on Pull Requests)
PyGithub

# For RenderAdminPower
requests>=2.25.0

# Add any additional dependencies for specific superpowers here
# For example, for web scraping:
# beautifulsoup4>=4.9.0
# requests>=2.25.0
# For data processing:
# pandas>=1.3.0
# numpy>=1.21.0